requirements.txt
streamlit
google-generativeaiss